# pgtype/hstore: Save 2 allocs in database/sql Scan implementation